backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
channels==4.2.0
channels-redis==4.2.1
langchain==0.3.20
langchain-openai==0.3.35
langchain-community==0.3.19
langgraph==0.6.9
mcp==1.15.0
langchain-mcp-adapters==0.1.11
python-dotenv==1.0.1
daphne==4.1.2
whitenoise==6.8.2
flower==2.0.1
pydantic==2.10.4
httpx==0.28.1
